refactor(wip): rename components to singular names
