+++ components/badge/_badges.scss
@@ -18,16 +18,16 @@ span.badge {
   &.new {
     font-weight: 300;
     font-size: 0.8rem;
-    color:  var(--md-sys-color-on-primary);
+    color: var(--md-sys-color-on-primary);
     background-color: var(--md-sys-color-primary);
     border-radius: 2px;
   }
   &.new:after {
-    content: " new";
+    content: ' new';
   }
 
   &[data-badge-caption]::after {
-    content: " " attr(data-badge-caption);
+    content: ' ' attr(data-badge-caption);
   }
 }
