+++ components/card/card.mjs
@@ -1,0 +1,10 @@
+import { Component } from '../atomic/component.mjs';
+
+class Card extends Component {
+  constructor(options) {
+    super(options);
+    this.addClassname('card');
+  }
+}
+
+export { Card };
