+++ components/list/_collection.scss
