+++ components/list/_list.scss
