+++ components/progress/_preloader.scss
