+++ components/slider/_range.scss
