+++ components/slider/range.ts
